Schedule and Task classes
Defined classes "Task" and "Schedule" for managing the actions of NPCs including changes due to possession.  Game-state-altering actions are stored individually as tasks and added to a schedule which has methods for updating the current task and making revisions based on player actions.  Also started implementing base scene for characters