Separate Fairy and NPC scenes, isPlayer parameter added to NPCs
made two different scenes for the fairy and for other characters.  This should make it easier to build different features into the two going forward.  NPCs now have a boolean parameter for if they are player-controlled currently, and will only respond to actions if they are player-controlled.  Also added joystick support