Possession and Unpossession
Functionality restored to before weird GitHub hiccup.  After possessing a character, the fairy's collision is disabled and the position relative to the possessed character is stored.  The Unpossess action releases the character, reactivates the fairy's collision, and places the fairy relative to the character based on the offset.  There is still a bug where if the fairy is in two interact regions when unpossession occurs the game crashes, will fix that next