Navigation almost works
NPC will move to desired positions in sequence if multiple tasks are added to their schedule.  However, if their task is just to stand in a location, they will be considered to have "failed" the task, regardless of if they reached the desired location or not.  Working on fixing this and testing other actions now