Technically a game
Pivoting direction a bit due to time constraints.  What we have now is playable and qualifies as a game, though it does not align strongly with the original vision.

The witch will follow the lumberjack and moves slightly faster than the lumberjack.  If the lumberjack picks up the axe, they can cut down trees to try to block the witch's path.  Survive as long as you can, but before you can control the lumberjack you must first learn their name.

The lumberjack's name is "Osama Hussein" (hinted to the player on the main menu video, can hint at it more in the page description if needed)